Remove the orphans handling from create table
We made DROP TABLE an atomic cluster state update a while ago, so there
shouldn't be cases anymore that can result in orphaned indices or
aliases.

Removing the logic that deleted those orphans on a CREATE TABLE should
therefore be safe. (And if not, it should expose any bugs and we can
then fix the root causes)